+++ docs/CHANGELOG.md
@@ -8,11 +8,10 @@ files use diff-like tools or "git log".
 
 ## Versions:
 ### v0.4.0 (WIP):
-- Rendering a whole line of typed chars/strings.
+- Rendering a whole lines of typed chars/strings.
 - UI redesigned.
 - Cursor added.
 - Makefile && similar ones improved.
-- Saving to a file.
 
 ### v0.3.0 (May 31, 2018):
 - Printing only pressed char on the screen.